Create budget
It’s your job to calculate the cost of replacing damaged battle droids and to check whether it is within the budget limit of Rs. 15000. The cost of the equipment and parts is given in the table below.

Blast Rifle

Rs. 350.34

Visual Sensors

Rs. 230.90

Auditory Sensors

Rs. 190.55

Arms

Rs. 125.30

Legs

Rs. 180.90

Can you write a program to do the given task?