Create control structure 1
Dev went a tour to Goa with his friends. There they went for boating. The rule is that only restricted weight can travel in a boat at a time so that the boat remains stable in the boat. You find that the boatman who is sailing your boat is so much greedy in making money. For earning more money he takes many number of peoples to travel in the boat at a time. So you wants to check how much number of people can travel in the boat at a time so that the boat will not drown. Calculate the weight by considering the number of adults and number of children.